+++ lib/misc/logger.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 void l(Object? message) {
   final isDebug = Zone.current[#debug] as bool? ?? false;
   if (isDebug) {
-    Zone.root.print(
+    Zone.current.print(
       '[$red${DateTime.now()}$reset]:$green $message $reset\n'
       '$yellow<=============================================>$reset',
     );