+++ lib/router/page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-typedef CustomBlazePageBuilder<T> = Route<T> Function(
+typedef CustomBlazePageBuilder<T extends Object> = Route<T> Function(
   BuildContext context,
 );
 
@@ -8,7 +8,7 @@ typedef BlazeDialogPageBuilder = Widget Function(
   BuildContext context,
 );
 
-class CustomBlazePage<T> extends Page<T> {
+class CustomBlazePage<T extends Object> extends Page<T> {
   const CustomBlazePage({required this.buildRoute});
 
   final CustomBlazePageBuilder<T> buildRoute;
@@ -17,7 +17,7 @@ class CustomBlazePage<T> extends Page<T> {
   Route<T> createRoute(BuildContext context) => buildRoute(context);
 }
 
-class BlazeDialogPage<T> extends Page<T> {
+class BlazeDialogPage<T extends Object> extends Page<T> {
   const BlazeDialogPage({required this.builder});
   final BlazeDialogPageBuilder builder;